tweaks to audio
* duration and position now take an return doubles
* sles was dealing with milliseconds instead of seconds..fixed

+++ include/XliAudio/SimpleAudio.h
@@ -10,14 +10,14 @@ namespace Xli
 	public:
 		virtual void Pause() = 0;
         virtual void UnPause() = 0;
-        virtual int GetPosition() const = 0;
-        virtual void SetPosition(int position) = 0;
+        virtual double GetPosition() const = 0;
+        virtual void SetPosition(double position) = 0;
 		virtual float GetVolume() const = 0;
 		virtual void SetVolume(float volume) const = 0;
 		virtual bool IsPlaying() = 0;
 		virtual bool IsFinished() = 0;
 
-        virtual int GetDuration() const = 0;
+        virtual double GetDuration() const = 0;
 		virtual float GetPan() const = 0;
 		virtual void SetPan(float pan) const = 0;
 		virtual void Play() = 0;
